requirement.txt

phidata
dotenv
groq
yfinance
